services/auth/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
mysqlclient==2.2.4
django-cors-headers==4.3.1 